Remove javaparser and fix broken tests from qdox migration.
