+++ metaserver/src/main/scala/scala/meta/languageserver/ctags/JavaCtags.scala
@@ -1,122 +1,118 @@
 package scala.meta.languageserver.ctags
 
-import scala.meta._
-import scala.meta.languageserver.ScalametaEnrichments._
-import com.github.javaparser.JavaParser
-import com.github.javaparser.ParseProblemException
-import com.github.javaparser.ParseStart.COMPILATION_UNIT
-import com.github.javaparser.ParserConfiguration
-import com.github.javaparser.Providers
-import com.github.javaparser.ast
-import com.github.javaparser.ast.Node
-import com.github.javaparser.ast.validator.ProblemReporter
-import com.github.javaparser.ast.validator.Validator
-import com.github.javaparser.ast.visitor.VoidVisitorAdapter
-import com.github.javaparser.{Position => JPosition}
+import java.io.StringReader
+import org.langmeta.inputs.Input
+import com.thoughtworks.qdox._
+import com.thoughtworks.qdox.model.JavaClass
 import org.langmeta.languageserver.InputEnrichments._
+import scala.meta.PACKAGE
+import scala.meta.TRAIT
+import scala.meta.OBJECT
+import scala.meta.CLASS
+import scala.meta.VAL
+import scala.meta.VAR
+import scala.meta.DEF
+import com.thoughtworks.qdox.model.JavaField
+import com.thoughtworks.qdox.model.JavaMember
+import com.thoughtworks.qdox.model.JavaMethod
+import com.thoughtworks.qdox.model.JavaModel
+import org.langmeta.inputs.Position
 
 object JavaCtags {
-  val parserConfig: ParserConfiguration = {
-    val config = new ParserConfiguration
-    // disable a few features we don't need, for performance reasons.
-    // early measurements with the default parser showed it handles ~10-15k loc/s,
-    // which is slow compared to 30-45k loc/s for scalameta.
-    config.setAttributeComments(false)
-    config.setDoNotConsiderAnnotationsAsNodeStartForCodeAttribution(true)
-    config.setValidator(new Validator {
-      override def accept(node: Node, problemReporter: ProblemReporter): Unit =
-        ()
-    })
-    config
+  private implicit class XtensionJavaModel(val m: JavaModel) extends AnyVal {
+    def lineNumber: Int = m.getLineNumber - 1
   }
-  def parseJavaCompilationUnit(contents: String): ast.CompilationUnit = {
-    val result = new JavaParser(parserConfig)
-      .parse(COMPILATION_UNIT, Providers.provider(contents))
-    if (result.isSuccessful) result.getResult.get
-    else throw new ParseProblemException(result.getProblems)
-  }
-
   def index(input: Input.VirtualFile): CtagsIndexer = {
-    def getPosition(t: ast.Node): Position.Range =
-      getPositionOption(t).getOrElse(Position.Range(input, -1, -1))
-    def getPositionOption(t: ast.Node): Option[Position.Range] =
-      for {
-        start <- Option(t.getBegin.orElse(null))
-        jend <- Option(t.getEnd.orElse(null))
-        end: JPosition = t match {
-          // Give names range positions.
-          case n: ast.expr.Name =>
-            new JPosition(start.line, start.column + n.asString().length)
-          case n: ast.expr.SimpleName =>
-            new JPosition(start.line, start.column + n.asString().length)
-          case _ => jend
-        }
-      } yield
-        input.toPosition(
-          start.line - 1,
-          start.column - 1,
-          end.line - 1,
-          end.column - 1
-        )
-    val cu: ast.CompilationUnit = parseJavaCompilationUnit(input.value)
-    new VoidVisitorAdapter[Unit] with CtagsIndexer {
-      override def language: String = "Java"
-      override def visit(
-          t: ast.PackageDeclaration,
-          ignore: Unit
-      ): Unit = {
-        val pos = getPosition(t.getName)
-        def loop(name: ast.expr.Name): Unit =
-          Option(name.getQualifier.orElse(null)) match {
-            case None =>
-              term(name.getIdentifier, pos, PACKAGE)
-            case Some(qual) =>
-              loop(qual)
-              term(name.getIdentifier, pos, PACKAGE)
+    val builder = new JavaProjectBuilder()
+    val source = builder.addSource(new StringReader(input.value))
+    source.getClasses.get(0).getNestedClasses
+    new CtagsIndexer { self =>
+      override def indexRoot(): Unit = {
+
+        /** Computes the start/end offsets from a name in a line number.
+         *
+         * Applies a simple heuristic to find the name: the first occurence of
+         * name in that line. If the name does not appear in the line then
+         * 0 is returned. If the name appears for example in the return type
+         * of a method then we get the position of the return type, not the
+         * end of the world.
+         */
+        def toRangePosition(line: Int, name: String): Position = {
+          val offset = input.toOffset(line, 0)
+          val column = {
+            val fromIndex = {
+              // Simple tricks to avoid hitting on keywords.
+              if (input.value.startsWith("package", offset))
+                "package".length
+              else if (input.value.startsWith("class", offset))
+                "class".length
+              else if (input.value.startsWith("interface", offset))
+                "interface".length
+              else if (input.value.startsWith("enum", offset))
+                "enum".length
+              else offset
+            }
+            val idx = input.value.indexOf(name, fromIndex)
+            if (idx == -1) 0
+            else idx - offset
           }
-        loop(t.getName)
-        super.visit(t, ignore)
+          val pos = input.toPosition(line, column, line, column + name.length)
+          pos
+        }
+
+        def visitFields[T <: JavaMember](fields: java.util.List[T]): Unit =
+          if (fields == null) ()
+          else fields.forEach(visitMember)
+
+        def visitClasses(classes: java.util.List[JavaClass]): Unit =
+          if (classes == null) ()
+          else classes.forEach(visitClass)
+
+        def visitClass(cls: JavaClass): Unit =
+          withOwner(owner(cls.isStatic)) {
+            val flags = if (cls.isInterface) TRAIT else CLASS
+            val pos = toRangePosition(cls.lineNumber, cls.getName)
+            withOwner() { term(cls.getName, pos, OBJECT) } // object
+            tpe(cls.getName, pos, flags)
+            visitClasses(cls.getNestedClasses)
+            visitFields(cls.getMethods)
+            visitFields(cls.getFields)
+            visitFields(cls.getEnumConstants)
+          }
+
+        def visitMember[T <: JavaMember](m: T): Unit =
+          withOwner(owner(m.isStatic)) {
+            val name = m.getName
+            val line = m match {
+              case c: JavaMethod => c.lineNumber
+              case c: JavaField => c.lineNumber
+              // TODO(olafur) handle constructos
+              case _ => 0
+            }
+            val pos = toRangePosition(line, name)
+            val flags: Long = m match {
+              case c: JavaMethod => DEF
+              case c: JavaField =>
+                if (c.isFinal) VAL
+                else VAR
+              // TODO(olafur) handle constructos
+              case _ => 0L
+            }
+            term(name, pos, flags)
+          }
+        if (source.getPackage != null) {
+          source.getPackageName.split("\\.").foreach { p =>
+            term(
+              p,
+              toRangePosition(source.getPackage.lineNumber, p),
+              PACKAGE
+            )
+          }
+        }
+        source.getClasses.forEach(visitClass)
       }
-      override def visit(
-          t: ast.body.EnumDeclaration,
-          ignore: Unit
-      ): Unit = withOwner(owner(t.isStatic)) {
-        term(t.getName.asString(), getPosition(t.getName), OBJECT)
-        super.visit(t, ignore)
-      }
-      override def visit(
-          t: ast.body.EnumConstantDeclaration,
-          ignore: Unit
-      ): Unit = withOwner() {
-        term(t.getName.asString(), getPosition(t.getName), VAL)
-        super.visit(t, ignore)
-      }
-      override def visit(
-          t: ast.body.ClassOrInterfaceDeclaration,
-          ignore: Unit
-      ): Unit = withOwner(owner(t.isStatic)) {
-        val name = t.getName.asString()
-        val pos = getPosition(t.getName)
-        // TODO(olafur) handle static methods/terms
-        if (t.isInterface) tpe(name, pos, TRAIT)
-        else tpe(name, pos, CLASS)
-        super.visit(t, ignore)
-      }
-      override def visit(
-          t: ast.body.FieldDeclaration,
-          ignore: Unit
-      ): Unit = withOwner(owner(t.isStatic)) {
-        val name = t.getVariables.get(0).getName
-        val flags = if (t.isFinal) VAL else VAR
-        term(name.asString(), getPosition(name), flags)
-      }
-      override def visit(
-          t: ast.body.MethodDeclaration,
-          ignore: Unit
-      ): Unit = withOwner(owner(t.isStatic)) {
-        term(t.getNameAsString, getPosition(t.getName), DEF)
-      }
-      override def indexRoot(): Unit = visit(cu, ())
+      override def language: String = "Java"
     }
   }
+
 }
